app/requirements.txt
flask
python-dotenv
requests
selenium